Revert "Refactor banner component and add lower banner component"
