+++ src/components/Banner.tsx
