+++ src/pages/home.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import MainLayout from '../layouts/MainLayout';
 import FeaturedProducts from '../components/Featured_Products';
-
-import LowerBanner from '../components/LowerBanner';
-import PImg from '../components/PImg';
+import Banner from '../components/Banner';
+import PromotionalImages from '../components/PromotionalImages'; // Use the new PromotionalImages component
 import BI from '../assets/banner_img.gif';
 import Breadcrumb from '../components/Breadcrumb';
 
@@ -19,10 +18,10 @@ const Home: React.FC = () => {
 
         {/* Featured Products and Banner */}
         <FeaturedProducts />
-        <LowerBanner 
-        imageUrl={BI}
-        altText="LowerBanner"
-      />
+        <Banner 
+          imageUrl={BI}
+          altText="Banner"
+        />
       </div>
     </MainLayout>
   );
