+++ src/pages/plp.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import "../styles/plp.css";
 import MainLayout from '../layouts/MainLayout';
-import LowerBanner from '../components/LowerBanner';
 import bannerImg from '../assets/banner-plp.jpg';
 import { ProductCard } from '../components/ProductCard';
 import { filtersData } from '../utils/dataFilter';
@@ -66,7 +65,6 @@ const Plp: React.FC<{ typeOfProduct: string }> = ({ typeOfProduct }) => {
                         <form>
                             <select
                                 id="orderby-box"
-                                title="SortOption"
                                 value={sortOption}
                                 onChange={(e) => setSortOption(e.target.value)} // Handle sorting option change
                             >
@@ -91,10 +89,7 @@ const Plp: React.FC<{ typeOfProduct: string }> = ({ typeOfProduct }) => {
                     </ul>
                 </section>
                 <div className="banner-plp">
-                <LowerBanner 
-                    imageUrl={bannerImg}
-                    altText="LowerBanner"
-                    />	
+                    <img src={bannerImg} alt="Banner" />
                 </div>
             </div>
         </MainLayout>
