+++ src/styles/plp.css
@@ -105,7 +105,7 @@ div.banner-plp{
 
 div.banner-plp img{
     display:block;
-    width: 100%;
+    width: 80%;
     height: auto;
     margin: 0;
     padding: 0;
